refactor(product-list): move into pages folder
