+++ src/app/pages/product/product.module.ts
@@ -1,18 +1,32 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { ReactiveFormsModule } from '@angular/forms';
+
+import { FlexLayoutModule } from '@angular/flex-layout';
+import { MaterialModule } from '@modules/material.module';
+import { NgxPaginationModule } from 'ngx-pagination';
 
 import { ProductRoutingModule } from './product-routing.module';
+
 import { ProductListComponent } from './product-list/product-list.component';
 import { ProductDetailComponent } from './product-detail/product-detail.component';
 import { ProductCreateComponent } from './product-create/product-create.component';
 import { ProductUpdateComponent } from './product-update/product-update.component';
 
-
 @NgModule({
-  declarations: [ProductListComponent, ProductDetailComponent, ProductCreateComponent, ProductUpdateComponent],
+  declarations: [
+    ProductListComponent,
+    ProductDetailComponent,
+    ProductCreateComponent,
+    ProductUpdateComponent,
+  ],
   imports: [
     CommonModule,
-    ProductRoutingModule
-  ]
+    ReactiveFormsModule,
+    ProductRoutingModule,
+    FlexLayoutModule,
+    MaterialModule,
+    NgxPaginationModule,
+  ],
 })
-export class ProductModule { }
+export class ProductModule {}
