--- src/app/product-list/product-list.component.html
@@ -1,112 +1,0 @@
-<div class="holder" fxLayout="row">
-  <div class="leftBlock" fxFlex></div>
-
-  <div class="container" fxFlex="90">
-    <form class="search-form" [formGroup]="form" (ngSubmit)="onSubmit()" fxLayout.gt-xs="row" fxLayout.lt-sm="column">
-      <div fxFlex="90"  class="inputStyle" fxLayout="row" fxLayout.xs="column" fxLayoutGap="0%" fxFlexFill>
-        <div fxFlex="100">
-          <div fxLayout="column" fxLayout.gt-xs="row" fxLayoutAlign="space-between">
-            <mat-form-field floatLabel="never" appearance="fill" fxFlex.gt-xs="37" fxFlex.lt-sm="100" (click)="picker.opened=true">
-              <mat-date-range-input [rangePicker]="picker" placeholder="日期">
-                <input matStartDate placeholder="Start date" formControlName="borrowStartDate">
-                <input matEndDate placeholder="End date" formControlName="borrowEndDate">
-              </mat-date-range-input>
-              <mat-datepicker-toggle matPrefix [for]="picker"></mat-datepicker-toggle>
-              <mat-date-range-picker #picker></mat-date-range-picker>
-            </mat-form-field>
-            <mat-form-field floatLabel="never" appearance="fill" fxFlex.gt-xs="33" fxFlex.lt-sm="100">
-              <i class="material-icons matPrefix-icon" matPrefix>location_on</i>
-              <mat-select formControlName="cityAreaName" placeholder="地區">
-                <mat-optgroup *ngFor="let cityName of city.nameArray; let cityIndex = index" [label]="cityName">
-                  <mat-option *ngFor="let areaName of city.areaNameArray[cityIndex]" [value]="areaName">
-                    {{areaName}}
-                  </mat-option>
-                </mat-optgroup>
-              </mat-select>
-            </mat-form-field>
-            <mat-form-field floatLabel="never" appearance="fill" fxFlex.gt-xs="30" fxFlex.lt-sm="100">
-              <i class="material-icons matPrefix-icon" matPrefix>local_atm</i>
-              <mat-select formControlName="priceRange" placeholder="預算">
-                <mat-option value="0">0 ~ 2,000</mat-option>
-                <mat-option value="1">2,001 ~ 4,000</mat-option>
-              </mat-select>
-            </mat-form-field>
-          </div>
-          <div>
-            <mat-form-field floatLabel="never" class="demo-chip-list" appearance="fill" fxFlex.gt-xs="100" fxFlex.lt-sm="100">
-              <mat-chip-list #chipList>
-                <mat-chip *ngFor="let type of chipTypes" [selectable]="selectable"
-                          [removable]="removable" (removed)="removeType(type)">
-                  {{type.name}}
-                  <mat-icon matChipRemove *ngIf="removable">cancel</mat-icon>
-                </mat-chip>
-                <input placeholder="New type..."
-                       #typeInput
-                       [matChipInputFor]="chipList"
-                       [matAutocomplete]="auto"
-                       [matChipInputSeparatorKeyCodes]="separatorKeysCodes"
-                       [matChipInputAddOnBlur]="addOnBlur"
-                       (matChipInputTokenEnd)="addType($event)">
-              </mat-chip-list>
-              <mat-autocomplete #auto="matAutocomplete" (optionSelected)="selected($event)">
-                <mat-option *ngFor="let type of productTypes" [value]="type.id">
-                  {{type.name}}
-                </mat-option>
-              </mat-autocomplete>
-            </mat-form-field>
-          </div>
-        </div>
-      </div>
-
-      <div fxFlex="10" fxFlex.lt-sm="100" class="search" fxLayoutAlign="center">
-        <button mat-button type="submit" id="search">
-          <i class="material-icons">search</i>
-        </button>
-      </div>
-    </form>
-
-    <div class="formTitle" fxLayout="row" fxLayoutAlign="left">
-      <h4>篩選結果</h4>
-    </div>
-
-    <div class="list">
-      <div fxLayout="row" fxLayout.lt-md="column" fxLayoutGap="10px">
-        <div fxLayout.gt-xs="row" fxLayout.lt-sm="column" fxLayoutAlign="space-between" fxLayoutGap="10px"
-          fxFlex="100" fxFlexFill>
-          <div fxFlex="30" fxFlex.lt-lg="50" fxFlex.lt-sm="100"
-            *ngFor="let productGroup of productGroups | paginate: { itemsPerPage: 10, currentPage: page }">
-            <mat-card class="card">
-              <div class="card-image">
-                <img mat-card-image [src]="productGroup.coverImage" alt="商品圖片">
-              </div>
-              <mat-card-content>
-                <h4 fxFlex.gt-md="66" fxFlex.lt-lg="100" fxFlex.sm="66">戶外野營折疊全自動帳篷</h4>
-                <div fxFlex.gt-md="30" fxFlex.lt-lg="100" fxFlex.sm="33" class="cost-and-score">
-                  <div fxLayoutWrap>
-                    <span class="cost">{{ productGroup.price }}</span>
-                    <p>
-                      {{ productGroup.city }}
-                      <span class="score">
-                        <i class="material-icons star" matPrefix>star_rate</i>5
-                      </span>
-                    </p>
-                  </div>
-                </div>
-              </mat-card-content>
-            </mat-card>
-          </div>
-        </div>
-      </div>
-    </div>
-
-    <div class="pagination">
-      <pagination-controls
-        (pageChange)="page = $event"
-        previousLabel=""
-        nextLabel=""
-      ></pagination-controls>
-    </div>
-  </div>
-
-  <div class="rightBlock" fxFlex></div>
-</div>
